Moved casting function from exchange to types
