+++ backend/libs/types/Cargo.toml
@@ -14,3 +14,4 @@ chrono = "0.4.19"
 warp = "0.3"
 tokio-tungstenite = { version = "0.13.0", features = ["tls"] }
 tokio-native-tls = "0.3"
+serde_json = "1.0.62"
