+++ backend/libs/types/src/errors.rs
@@ -3,6 +3,31 @@ use ::std::fmt::{Display, Formatter, Result as FormatResult};
 
 use ::serde::Serialize;
 
+#[derive(Debug, Clone)]
+pub struct ParseError {
+  fld_name: String,
+}
+
+impl ParseError {
+  pub fn new(fld_name: &str) -> Self {
+    return ParseError {
+      fld_name: fld_name.into(),
+    };
+  }
+}
+
+impl Display for ParseError {
+  fn fmt(&self, f: &mut Formatter<'_>) -> FormatResult {
+    return write!(f, "Failed to parse '{}", self.fld_name);
+  }
+}
+
+impl Error for ParseError {
+  fn source(&self) -> Option<&(dyn Error + 'static)> {
+    None
+  }
+}
+
 #[derive(Debug, Clone, Serialize)]
 pub struct VecElementErr<T>
 where
